requirements.txt
# Core PDF processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# OCR
pytesseract>=0.3.10
# easyocr>=1.7.0  # optional: pip install easyocr (large dependency)

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Output generation
openpyxl>=3.1.2
pandas>=2.0.0

# Numeric
numpy>=1.24.0

# External system requirements (install separately):
# - Tesseract OCR: https://github.com/tesseract-ocr/tesseract
#   Windows: https://github.com/UB-Mannheim/tesseract/wiki
#   macOS:   brew install tesseract tesseract-lang
#   Linux:   sudo apt install tesseract-ocr tesseract-ocr-kor tesseract-ocr-eng
